+++ packages/telescope-core/lib/client/templates/common/checker.js
@@ -1,0 +1,5 @@
+Template.checker.helpers({
+  allow: function () {
+    return Users.can[this.check](Meteor.user(), this.doc);
+  }
+});