+++ packages/telescope-core/lib/client/templates/errors/no_rights.js
@@ -1,6 +1,5 @@
 Template.no_rights.helpers({
   errorMessage: function () {
-    console.log(this)
     return !!this.message ? i18n.t(this.message) : i18n.t("sorry_you_dont_have_the_rights_to_view_this_page");
   }
 });