+++ packages/telescope-lib/lib/router.js
@@ -13,3 +13,18 @@ FlowRouter.removeFromQueryArray = function (key, value) {
   params[key] = keyArray;
   FlowRouter.setQueryParams(params);
 }
+
+Telescope.adminRoutes = FlowRouter.group({
+  prefix: '/admin',
+  name: 'admin'
+});
+
+FlowRouter.notFound = {
+  action: function() {
+    BlazeLayout.render("layout", {main: "not_found"});
+  }
+};
+
+FlowRouter.triggers.exit([Messages.clearSeen]);
+FlowRouter.triggers.exit([Events.analyticsInit]); // will only run once thanks to _.once()
+FlowRouter.triggers.exit([Events.analyticsRequest]);