+++ packages/telescope-lib/lib/router/admin.js
@@ -1,4 +1,0 @@
-Telescope.adminRoutes = FlowRouter.group({
-  prefix: '/admin',
-  name: 'admin'
-});