+++ packages/telescope-messages/lib/client/messages.js
@@ -9,6 +9,7 @@ Messages = {
   },
 
   clearSeen: function() {
+    console.log("clearing messages…")
     this.collection.update({seen:true}, {$set: {show:false}}, {multi:true});
   }
 };