+++ packages/telescope-users/lib/permissions.js
@@ -1,3 +1,6 @@
+// note: using collection helpers here is probably a bad idea, 
+// because they'll throw an error when the user is undefined
+
 /**
  * Telescope permissions
  * @namespace Users.can
@@ -20,6 +23,8 @@ Users.can.view = function (user) {
   }
   return true;
 };
+Users.helpers({canView: function () {return Users.can.view(this);}});
+
 
 Users.can.viewById = function (userId) {
   // if an invite is required to view, run permission check, else return true
@@ -28,6 +33,7 @@ Users.can.viewById = function (userId) {
   }
   return true;
 };
+Users.helpers({canViewById: function () {return Users.can.viewById(this);}});
 
 Users.can.viewPendingPosts = function (user) {
   user = (typeof user === 'undefined') ? Meteor.user() : user;
@@ -37,7 +43,6 @@ Users.can.viewPendingPosts = function (user) {
 Users.can.viewPendingPost = function (user, post) {
   return Users.is.owner(user, post) || Users.can.viewPendingPosts(user);
 };
-
 
 Users.can.viewRejectedPosts = function (user) {
   user = (typeof user === 'undefined') ? Meteor.user() : user;
@@ -65,14 +70,17 @@ Users.can.post = function (user, returnError) {
     return true;
   }
 };
+Users.helpers({canPost: function () {return Users.can.post(this);}});
 
 Users.can.comment = function (user, returnError) {
   return Users.can.post(user, returnError);
 };
+Users.helpers({canComment: function () {return Users.can.comment(this);}});
 
 Users.can.vote = function (user, returnError) {
   return Users.can.post(user, returnError);
 };
+Users.helpers({canVote: function () {return Users.can.vote(this);}});
 
 /**
  * Check if a user can edit a document
@@ -91,11 +99,13 @@ Users.can.edit = function (user, document) {
 
   return adminCheck || ownerCheck;
 };
+Users.helpers({canEdit: function (document) {return Users.can.edit(this, document);}});
 
 Users.can.editById = function (userId, document) {
   var user = Meteor.users.findOne(userId);
   return Users.can.edit(user, document);
 };
+Users.helpers({canEditById: function (document) {return Users.can.editById(this, document);}});
 
 /**
  * Check if a user can submit a field
@@ -114,6 +124,7 @@ Users.can.submitField = function (user, field) {
   return adminCheck || memberCheck;
 
 };
+Users.helpers({canSubmitField: function (field) {return Users.can.submitField(this, field);}});
 
 /** @function
  * Check if a user can edit a field – for now, identical to Users.can.submitField
@@ -125,3 +136,5 @@ Users.can.editField = Users.can.submitField;
 Users.can.invite = function (user) {
   return Users.is.invited(user) || Users.is.admin(user);
 };
+Users.helpers({canInvite: function () {return Users.can.invite(this);}});
+